♻️Refactor : ProductViewController 구현 및 상속
