+++ OpenMarket/OpenMarket/ProductViewController.swift
@@ -1,0 +1,75 @@
+//
+//  ProductViewController.swift
+//  OpenMarket
+//
+//  Created by Grumpy, OneTool on 2022/05/26.
+//
+
+import UIKit
+
+class ProductViewController: UIViewController {
+    lazy var productView = ProductView(frame: view.frame)
+    weak var delegate: ListUpdatable?
+    var currency: Currency = .KRW
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        self.view = productView
+        self.view.backgroundColor = .white
+        defineTextFieldDelegate()
+        productView.currencyField.addTarget(self, action: #selector(changeCurrency(_:)), for: .valueChanged)
+        productView.currencyField.selectedSegmentIndex = currency.value
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification , object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide), name: UIResponder.keyboardWillHideNotification, object: nil)
+        
+    }
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification , object: nil)
+        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
+    }
+    
+    @objc func keyboardWillShow(notification: NSNotification) {
+        if let keyboardSize = (notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue)?.cgRectValue {
+            self.productView.mainScrollView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: keyboardSize.height, right: 0)
+            
+            if self.productView.descriptionView.isFirstResponder {
+                productView.mainScrollView.scrollRectToVisible(productView.descriptionView.frame, animated: true)
+            }
+        }
+    }
+    
+    @objc func keyboardWillHide(notification: NSNotification) {
+        self.productView.mainScrollView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: 10, right: 0)
+    }
+    
+    @objc func changeCurrency(_ sender: UISegmentedControl) {
+        let mode = sender.selectedSegmentIndex
+        if mode == Currency.KRW.value {
+            currency = Currency.KRW
+        } else if mode == Currency.USD.value {
+            currency = Currency.USD
+        }
+    }
+    
+    private func defineTextFieldDelegate() {
+        productView.priceField.delegate = self
+        productView.discountedPriceField.delegate = self
+        productView.stockField.delegate = self
+    }
+}
+
+extension ProductViewController: UITextFieldDelegate {
+    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
+        if textField.keyboardType == .numberPad {
+            if CharacterSet(charactersIn: "0123456789").isSuperset(of: CharacterSet(charactersIn: string)) {
+                return true
+            }
+        }
+        return false
+    }
+}
